requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-docx==1.1.2
pdfminer.six==20231228
numpy==1.26.4
scikit-learn==1.5.2
sentence-transformers==3.0.1
spacy==3.7.6
requests==2.32.3
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.3.3
